게시글 수정, 삭제 요청자가 게시글 작성자인 경우에만 메소드 실행되도록 처리, User의 id Long -> int로 변경

+++ src/main/java/com/example/last/entity/User.java
@@ -14,7 +14,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private String email;
@@ -35,7 +35,7 @@ public class User {
     public void setPassword(String password) { this.password = password; }
 
     @Builder
-    public User(Long id, String email, String password, String nickname, Authority authority) {
+    public User(Integer id, String email, String password, String nickname, Authority authority) {
         this.id = id;
         this.email = email;
         this.password = password;
